refactor(greatestCommonDivisor)
* added readme file
* fixed implementation
* added respective test case
