+++ recursion/greatestCommonDivison/readme.md
@@ -1,0 +1,10 @@
+# greatestCommonDivisor
+
+---
+
+Write recursive `greatestCommonDivision` function computing the greatest common divisor of two positive numbers
+
+```js
+greatestCommonDivisor(2154, 458); // 2
+greatestCommonDivisor(458, 2154); // 2
+```
